// src/app/page/master/pegawai-sk-gaji-bak/pegawai-sk-gaji-bak.interface.ts
import { EventEmitter } from '@angular/core';

export interface PegawaiSkGajiBAK {
    kode?: any;
    pegawai?: number;
    hargaSatuan?: string
    kdAkunBank?: string;
    kdKomponenHarga?: string;
    persenHargaSatuan?: string;
    status1?: string;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}